+++ CHANGELOG.md
@@ -18,6 +18,8 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
   $env:Temp and reused every time DSC check runs, instead of creating a new
   module every time.
   - Added AD Permissions parameter for xExchReceiveConnector.
+  - xExchAddressList: Removing the scriptblock creation for RecipientFilter
+  property in Get-TargetResource.
   - Adding missing TlsAuthLevel to xExchSendConnector Get-TargetResource function. 
   - Small bug fix in xExchangeHelper module.